refactor: Table 서비스 Order import 제거
