+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,12 +1,10 @@
 package kitchenpos.table.application;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.dao.OrderDao;
 import kitchenpos.table.dao.OrderTableDao;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableValidator;
 import kitchenpos.table.dto.request.OrderTableCreateRequest;
 import kitchenpos.table.dto.response.OrderTableResponse;
 import org.springframework.stereotype.Service;
@@ -15,12 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class TableService {
 
-    private final OrderDao orderDao;
     private final OrderTableDao orderTableDao;
+    private final TableValidator tableValidator;
 
-    public TableService(final OrderDao orderDao, final OrderTableDao orderTableDao) {
-        this.orderDao = orderDao;
+    public TableService(final OrderTableDao orderTableDao, final TableValidator tableValidator) {
         this.orderTableDao = orderTableDao;
+        this.tableValidator = tableValidator;
     }
 
     @Transactional
@@ -38,10 +36,8 @@ public class TableService {
     @Transactional
     public OrderTableResponse changeEmpty(final Long orderTableId, final boolean empty) {
         final OrderTable savedOrderTable = findOrderTable(orderTableId);
-        validateTableOrderStatus(orderTableId);
-
+        tableValidator.validateTable(orderTableId);
         savedOrderTable.changeEmpty(empty);
-
         final OrderTable orderTable = orderTableDao.save(savedOrderTable);
         return OrderTableResponse.from(orderTable);
     }
@@ -49,13 +45,6 @@ public class TableService {
     private OrderTable findOrderTable(final Long orderTableId) {
         return orderTableDao.findById(orderTableId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 주문 테이블입니다."));
-    }
-
-    private void validateTableOrderStatus(final Long orderTableId) {
-        if (orderDao.existsByOrderTableIdAndOrderStatusIn(
-                orderTableId, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
-            throw new IllegalArgumentException("주문 상태가 Cooking이나 Meal일 경우 테이블의 상태를 변경할 수 없습니다.");
-        }
     }
 
     @Transactional
