+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -13,14 +13,15 @@ import kitchenpos.dao.FakeOrderDao;
 import kitchenpos.dao.FakeOrderTableDao;
 import kitchenpos.dao.FakeTableGroupDao;
 import kitchenpos.order.dao.OrderDao;
+import kitchenpos.table.application.TableGroupService;
 import kitchenpos.table.dao.OrderTableDao;
 import kitchenpos.table.dao.TableGroupDao;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.domain.TableValidator;
 import kitchenpos.table.dto.request.OrderTableRequest;
 import kitchenpos.table.dto.request.TableGroupRequest;
 import kitchenpos.table.dto.response.TableGroupResponse;
-import kitchenpos.table.application.TableGroupService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -32,13 +33,15 @@ public class TableGroupServiceTest {
     private OrderDao orderDao;
     private OrderTableDao orderTableDao;
     private TableGroupDao tableGroupDao;
+    private TableValidator tableValidator;
 
     @BeforeEach
     void beforeEach() {
         this.orderDao = new FakeOrderDao();
         this.orderTableDao = new FakeOrderTableDao();
         this.tableGroupDao = new FakeTableGroupDao();
-        this.tableGroupService = new TableGroupService(orderDao, orderTableDao, tableGroupDao);
+        this.tableValidator = new TableValidator(orderDao);
+        this.tableGroupService = new TableGroupService(orderTableDao, tableGroupDao, tableValidator);
     }
 
     @Test
